requirements.txt
# MCP Python SDK
mcp[cli]>=1.11.0

# Game automation dependencies
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
mss>=9.0.1

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
asyncio-throttle>=1.0.2
